+++ corva/__init__.py
@@ -1,7 +1,7 @@
 from .application import Corva
 from .models.scheduled import ScheduledEvent
 from .models.stream import StreamEvent
-from .network.api import Api
+from .api import Api
 from .state.redis_state import RedisState as Cache
 
 __all__ = [
